rebalance crypto calculations, consume a lot more power remove emag and emp acts add liquid coolant to replace atmos molar factor, new reagent add app interface to admin buttons for convenience machines don't spam chat on payouts events rebalanced, removed algo change code cleanup
