+++ RussStation.dme
@@ -4752,6 +4752,7 @@
 #include "russstation\code\game\turfs\simulated\minerals.dm"
 #include "russstation\code\modules\admin\smites\confetti.dm"
 #include "russstation\code\modules\admin\smites\confetti_gib.dm"
+#include "russstation\code\modules\admin\verbs\cryptocurrency.dm"
 #include "russstation\code\modules\admin\verbs\randomverbs.dm"
 #include "russstation\code\modules\antagonists\ert.dm"
 #include "russstation\code\modules\antagonists\nukeop\equipment\pinpointer.dm"
