+++ russstation/code/modules/reagents/chemistry/recipes/others.dm
@@ -5,3 +5,9 @@
 /datum/chemical_reaction/russ/confettium
 	results = list(/datum/reagent/confettium = 1)
 	required_reagents = list(/datum/reagent/consumable/laughter = 1, /datum/reagent/lube = 1, /datum/reagent/colorful_reagent = 1)
+
+// proprietary formula donut steel
+/datum/chemical_reaction/russ/super_coolant
+	results = list(/datum/reagent/super_coolant = 1)
+	required_reagents = list(/datum/reagent/cryostylane = 1, /datum/reagent/water/hollowwater = 1)
+	required_catalysts = list(/datum/reagent/bluespace = 1)
